64 bit rdrand functions in x86_64, 16 and 32 in i686
